[IMP] html_editor: rm MERGE_ADJACENT_NODES and CLEAN on save
This commit separates the concerns of the CLEAN and CLEAN_FOR_SAVE
commands.

CLEAN_FOR_SAVE, as the name implies, is dispatched when about to save
the editor's content. It is meant to be called on a clone of the
editor's content (except when testing with `testEditor`).

CLEAN is dispatched by plugins when they need to clean a subtree before
they perform an operation on it, for example removing the hint-related
attributes before cloning a element. The changes made by the CLEAN
handlers are typically reversed by the NORMALIZE handlers.

This commit, other than redefining how certain plugins handle CLEAN and
CLEAN_FOR_SAVE, removes the call to CLEAN from the editor's
`getElContent` method. It also removes the call to MERGE_ADJACENT_NODES.
If fact, it removes the MERGE_ADJACENT_NODES command altogether, and
makes such operation part of the NORMALIZE and CLEAN_FOR_SAVE handlers
in the format plugin.
