+++ addons/html_editor/static/src/editor.js
@@ -215,8 +215,6 @@ export class Editor {
 
     getElContent() {
         const el = this.editable.cloneNode(true);
-        this.dispatch("CLEAN", { root: el });
-        this.dispatch("MERGE_ADJACENT_NODE", { node: el });
         this.dispatch("CLEAN_FOR_SAVE", { root: el });
         return el;
     }
