+++ addons/html_editor/static/src/others/qweb_plugin.js
@@ -47,17 +47,10 @@ export class QWebPlugin extends Plugin {
                 this.normalize(payload.node);
                 break;
             case "CLEAN":
-                // TODO @phoenix: evaluate if this should be cleanforsave instead
-                for (const node of payload.root.querySelectorAll(
-                    "[data-oe-t-group], [data-oe-t-inline], [data-oe-t-selectable], [data-oe-t-group-active]"
-                )) {
-                    node.removeAttribute("data-oe-t-group-active");
-                    node.removeAttribute("data-oe-t-group");
-                    node.removeAttribute("data-oe-t-inline");
-                    node.removeAttribute("data-oe-t-selectable");
-                }
+                this.clearDataAttributes(payload.root);
                 break;
             case "CLEAN_FOR_SAVE":
+                this.clearDataAttributes(payload.root);
                 for (const element of payload.root.querySelectorAll(
                     "[t-esc], [t-raw], [t-out], [t-field]"
                 )) {
@@ -203,5 +196,16 @@ export class QWebPlugin extends Plugin {
         this.selectedNode = node;
         this.picker.close();
         this.selectNode(node);
+    }
+
+    clearDataAttributes(root) {
+        for (const node of root.querySelectorAll(
+            "[data-oe-t-group], [data-oe-t-inline], [data-oe-t-selectable], [data-oe-t-group-active]"
+        )) {
+            node.removeAttribute("data-oe-t-group-active");
+            node.removeAttribute("data-oe-t-group");
+            node.removeAttribute("data-oe-t-inline");
+            node.removeAttribute("data-oe-t-selectable");
+        }
     }
 }
